Make push_apk and push_aab async functions
When I added support for the samsung galaxy store, I used aiohttp
because I didn't pay much attention to the fact that this lib was sync
when scriptworker is async. I worked around the sync -> async barrier in
`push_apk` by starting a new event loop and running the async code in
there. In theory this works. In practice though, I forgot that asyncio
does **not** like nesting event loops and it will throw exceptions if
you try... So let's make this library async too, it's a very minimal
change for any user. Either they're already in async land and they just
have to slap an `await` in front of the call (and enjoy the fact that it
doesn't traceback anymore) or they're in sync land and they can slap an
`asyncio.run` around it.

Note that `push_aab` does not need to be async but I marked it as such
for consistency between the 2 APIs.