Always extract architecture metadata from APKs (#283)
* Always extract architecture metadata from APKs

We depend on the architecture being present in the metadata when
uploading binaries to the samsung galaxy store. This isn't the case for
focus because we pass `skip_check_ordered_version_codes` which in turn
leads to the architecture metadata not being extracted.

Unlike the other flagged metadata extraction, this one doesn't seem to
be able to fail nor does it make this function slower (there's less than
a 1% perf difference) so let's just unconditionally extract it and
remove part of a footgun in the API.

Fixes #282

* Add tests for other extract_metadata parameter permutations